add instruments
