+++ src/main/java/de/ftscraft/ftsengine/feature/instruments/CustomInstrument.java
@@ -1,0 +1,14 @@
+package de.ftscraft.ftsengine.feature.instruments;
+
+import org.bukkit.inventory.Inventory;
+import org.jetbrains.annotations.NotNull;
+
+public class CustomInstrument implements Instrument {
+
+
+
+    @Override
+    public @NotNull Inventory getInventory() {
+        return null;
+    }
+}
